yolo/consumer/yolo_inference/requirements.txt
confluent-kafka
Pillow

#torch
ultralytics
